requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
djangorestframework==3.10.3
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
oauth2client==4.1.3
Pillow==6.2.0
psycopg2-binary==2.7.5
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-decouple==3.1
pytz==2019.2
razorpay==1.2.0
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==3.3.0
